# Rename fromAdjacencyList to stars