--- a/src/Algebra/Graph.hs
+++ b/src/Algebra/Graph.hs
@@ -24,7 +24,6 @@ module Algebra.Graph (
 
     -- * Basic graph construction primitives
     empty, vertex, edge, overlay, connect, vertices, edges, overlays, connects,
-    fromAdjacencyList,
 
     -- * Graph folding
     foldg,
@@ -37,8 +36,8 @@ module Algebra.Graph (
     edgeList, vertexSet, vertexIntSet, edgeSet, adjacencyList,
 
     -- * Standard families of graphs
-    path, circuit, clique, biclique, star, starTranspose, tree, forest, mesh,
-    torus, deBruijn,
+    path, circuit, clique, biclique, star, stars, starTranspose, tree, forest,
+    mesh, torus, deBruijn,
 
     -- * Graph transformation
     removeVertex, removeEdge, replaceVertex, mergeVertices, splitVertex,
@@ -334,19 +333,6 @@ overlays = foldr overlay empty
 connects :: [Graph a] -> Graph a
 connects = foldr connect empty
 
--- | Construct a graph from an adjacency list.
--- Complexity: /O((n + m))/ time, memory and size.
---
--- @
--- fromAdjacencyList []                                  == 'empty'
--- fromAdjacencyList [(x, [])]                           == 'vertex' x
--- fromAdjacencyList [(x, [y])]                          == 'edge' x y
--- fromAdjacencyList . 'adjacencyList'                     == id
--- 'overlay' (fromAdjacencyList xs) (fromAdjacencyList ys) == fromAdjacencyList (xs ++ ys)
--- @
-fromAdjacencyList :: [(a, [a])] -> Graph a
-fromAdjacencyList = overlays . map (uncurry star)
-
 -- | Generalised 'Graph' folding: recursively collapse a 'Graph' by applying
 -- the provided functions to the leaves and internal nodes of the expression.
 -- The order of arguments is: empty, vertex, overlay and connect.
@@ -591,7 +577,7 @@ edgeIntSet = AIM.edgeSet . fromGraphAIM
 -- adjacencyList ('vertex' x)          == [(x, [])]
 -- adjacencyList ('edge' 1 2)          == [(1, [2]), (2, [])]
 -- adjacencyList ('star' 2 [3,1])      == [(1, []), (2, [1,3]), (3, [])]
--- 'fromAdjacencyList' . adjacencyList == id
+-- 'stars' . adjacencyList == id
 -- @
 {-# SPECIALISE adjacencyList :: Graph Int -> [(Int,[Int])] #-}
 adjacencyList :: Ord a => Graph a -> [(a, [a])]
@@ -693,6 +679,23 @@ biclique xs ys = connect (vertices xs) (vertices ys)
 star :: a -> [a] -> Graph a
 star x [] = vertex x
 star x ys = connect (vertex x) (vertices ys)
+
+-- | The /stars/ formed by overlaying a list of 'star's. An inverse of
+-- 'adjacencyList'.
+-- Complexity: /O(L)/ time, memory and size, where /L/ is the total size of the
+-- input.
+--
+-- @
+-- stars []                      == 'empty'
+-- stars [(x, [])]               == 'vertex' x
+-- stars [(x, [y])]              == 'edge' x y
+-- stars [(x, ys)]               == 'star' x ys
+-- stars                         == 'overlays' . map (uncurry 'star')
+-- stars . 'adjacencyList'         == id
+-- 'overlay' (stars xs) (stars ys) == stars (xs ++ ys)
+-- @
+stars :: [(a, [a])] -> Graph a
+stars = overlays . map (uncurry star)
 
 -- | The /star transpose/ formed by a list of leaves connected to a centre vertex.
 -- Complexity: /O(L)/ time, memory and size, where /L/ is the length of the