--- a/src/Algebra/Graph.hs
+++ b/src/Algebra/Graph.hs
@@ -573,11 +573,11 @@ edgeIntSet = AIM.edgeSet . fromGraphAIM
 -- Complexity: /O(n + m)/ time and /O(m)/ memory.
 --
 -- @
--- adjacencyList 'empty'               == []
--- adjacencyList ('vertex' x)          == [(x, [])]
--- adjacencyList ('edge' 1 2)          == [(1, [2]), (2, [])]
--- adjacencyList ('star' 2 [3,1])      == [(1, []), (2, [1,3]), (3, [])]
--- 'stars' . adjacencyList == id
+-- adjacencyList 'empty'          == []
+-- adjacencyList ('vertex' x)     == [(x, [])]
+-- adjacencyList ('edge' 1 2)     == [(1, [2]), (2, [])]
+-- adjacencyList ('star' 2 [3,1]) == [(1, []), (2, [1,3]), (3, [])]
+-- 'stars' . adjacencyList        == id
 -- @
 {-# SPECIALISE adjacencyList :: Graph Int -> [(Int,[Int])] #-}
 adjacencyList :: Ord a => Graph a -> [(a, [a])]