--- a/src/Algebra/Graph/Relation/Internal.hs
+++ b/src/Algebra/Graph/Relation/Internal.hs
@@ -188,14 +188,14 @@ instance (Ord a, Num a) => Num (Relation a) where
 -- /Note: this function is for internal use only/.
 --
 -- @
--- consistent 'Algebra.Graph.Relation.empty'                  == True
--- consistent ('Algebra.Graph.Relation.vertex' x)             == True
--- consistent ('Algebra.Graph.Relation.overlay' x y)          == True
--- consistent ('Algebra.Graph.Relation.connect' x y)          == True
--- consistent ('Algebra.Graph.Relation.edge' x y)             == True
--- consistent ('Algebra.Graph.Relation.edges' xs)             == True
--- consistent ('Algebra.Graph.Relation.graph' xs ys)          == True
--- consistent ('Algebra.Graph.Relation.fromAdjacencyList' xs) == True
+-- consistent 'Algebra.Graph.Relation.empty'         == True
+-- consistent ('Algebra.Graph.Relation.vertex' x)    == True
+-- consistent ('Algebra.Graph.Relation.overlay' x y) == True
+-- consistent ('Algebra.Graph.Relation.connect' x y) == True
+-- consistent ('Algebra.Graph.Relation.edge' x y)    == True
+-- consistent ('Algebra.Graph.Relation.edges' xs)    == True
+-- consistent ('Algebra.Graph.Relation.graph' xs ys) == True
+-- consistent ('Algebra.Graph.Relation.stars' xs)    == True
 -- @
 consistent :: Ord a => Relation a -> Bool
 consistent (Relation d r) = referredToVertexSet r `Set.isSubsetOf` d