… CHANGES.md
… test/Algebra/Graph/Test/API.hs
… test/Algebra/Graph/Test/Generic.hs
… test/Algebra/Graph/Test/NonEmptyGraph.hs
… test/Algebra/Graph/Test/Relation.hs