See https://github.com/snowleopard/alga/issues/99#issuecomment-408714915