@snowleopard 2018-07-30
@nobrakal One thing that bothers me is `starTranspose` and possible `starsTranspose` which may similarly be useful. It feels that the API shouldn't contain them, because `transpose . stars` does the same thing as `starsTranspose`, but unfortunately `transpose . stars` would be slower. I wonder if there is a way to fuse `transpose` with other graph construction functions nicely?

For example, we could have a rule `transpose (connect x y) = connect (transpose y) (transpose x)`, as well as `transpose . vertex = vertex`. Could you investigate this?