@nobrakal 2018-07-30
> I wonder if there is a way to fuse transpose with other graph construction functions nicely?

Yep, this is possible using rewrite rules :)

```Haskell
{-# RULES
"transpose/connect"  forall g1 g2. transpose (connect g1 g2) = connect (transpose g2) (transpose g1)
"transpose/vertex"   forall v. transpose (vertex v) = vertex v
"transpose/vertices" forall xs. transpose (vertices xs) = vertices xs
"transpose/stars" forall xs. transpose (stars xs) = overlays (map (transpose . uncurry star) xs)
 #-}
```

(I have also added some `INLINE` and `NOOINLINE` pragmas for concerned functions).

Sadly, all of this includes `overlays` which is very particular: due to its simplicity, adding any `INLINE` pragma drops its performances (and I don't know why). Butif I want my rule to fire, I have to delay `overlays`'s inlining...

So I had to trick the thing using less generic rules.

Anyway the result is here:

Using:

```Haskell
import Algebra.Graph

import Criterion.Main

import Control.DeepSeq

main :: IO ()
main = defaultMain
  [
    bench "starTranspose" $ nf (starTranspose 0) $!! intList
  , bench "star . transpose" $ nf (transpose . star 0) $!! intList
  , bench "starsTranspose" $ nf starsTranspose $!! toStars
  , bench "transpose . stars" $ nf starsTranspose' $!! toStars
  ]

starsTranspose :: [(a,[a])] -> Graph a
starsTranspose = overlays . map (uncurry starTranspose)

starsTranspose' :: [(a,[a])] -> Graph a
starsTranspose' = transpose . stars

toStars :: [(Int,[Int])]
toStars = map (\x -> (x,intList)) intList

intList :: [Int]
intList = [1..1000]
```

With `-O2`:

```
benchmarking starTranspose
time                 19.75 μs   (19.67 μs .. 19.85 μs)
[...]

benchmarking star . transpose
time                 19.83 μs   (19.78 μs .. 19.88 μs)
[...]

benchmarking starsTranspose
time                 20.42 ms   (20.31 ms .. 20.60 ms)
[...]

benchmarking transpose . stars
time                 20.43 ms   (20.27 ms .. 20.74 ms)
[...]
```

And using `-O2 -fno-enable-rewrite-rules`:
```
benchmarking starTranspose
time                 19.71 μs   (19.61 μs .. 19.82 μs)
[...]

benchmarking star . transpose
time                 39.91 μs   (39.13 μs .. 40.71 μs)
[...]

benchmarking starsTranspose
time                 21.23 ms   (20.72 ms .. 21.88 ms)
[...]

benchmarking transpose . stars
time                 39.41 ms   (38.30 ms .. 40.93 ms)
[...]
```