@snowleopard 2018-07-30
@nobrakal Cool, thank you! Presumably, you don't necessarilty need to use `-O2`? Will `-O` be sufficient?

> Sadly, all of this includes `overlays` which is very particular: due to its simplicity, adding any INLINE pragma drops its performances (and I don't know why).

Yes, we need to figure out what to do with `overlays`. Anyway, thanks for the useful experiment -- I will remove `starTranspose` and we can add your rewrite rules later, in a separate PR.