@nobrakal 2018-07-30
> Presumably, you don't necessarilty need to use -O2? Will -O be sufficient?

Yep, I only  used`-O2` because it is the flag of alga build :)

Ok, prefect, I just want to check what happen to `overlays` and try to find better rules before