@snowleopard 2018-07-30
> Yep, I only used `-O2` because it is the flag of alga build :)

Actually, let me try to drop it -- see #101.